v0.34
This is a somewhat experimental update.
Changes:
*) The sieve prime limit is now a separate parameter
*) pps display removed, instead we have sieve primes to non-primes ratio
and a badly estimated chance to find a block (chance excludes fractional
difficulty)
*) The miner searches only for Cunningham chains per default now. Twin
chain search is optional and will make sieving a lot slower but
increases the chance to actually find a twin chain. Twin chains can
still be found 'by accident' however