More parameters for tweaking
Add -d (SievePercentage) & -primes (SievePrimeLimit) command line
parameters that can be used for more tweaking
SievePercentage  ranges from 1 - 100, recommended is about 10 (default
is 8). It limits how many base primes are used to filter out candidate
multipliers in the sieve

The idea of SievePrimeLimit was taken from the latest official
jhPrimeMiner source. The default value is equal to MaxSieveSize which I
still recommend.
Changed the sampling time for the multiplier auto adjustment to 10 sec.