New proof of work for pool mining
Changed the way how block data is handled. I also added limits for all
parameters that affect the result of the sieve (the limits are sent by
the server together with block data). Additionally, the miner now
collects info about the mining process and sends it to the server about
once a minute. This information is used to validate and reward the
miners work instead of relying on shares.